alx_travel_app/requirements.txt
Django==4.2.24
djangorestframework==3.16.1
django-cors-headers==4.7.0
celery==5.5.3
redis==6.4.0
django-celery-results==2.6.0
django-celery-beat==2.8.1
drf-yasg==1.21.10
django-environ==0.12.0
mysqlclient==2.2.7
